+++ DataAccess/DataAccess/QueryServices/GetArticleByIdQueryService.cs
@@ -33,7 +33,7 @@ namespace DataAccess.QueryServices
             return Database.TryExecuteTransaction((transaction) =>
             {
                 var dbQuery = Database.CreateStoredProcCommand("Blog.GetArticleById", transaction);
-                var parameter = Database.CreateParameter("ArticleId", query.ArticleID);
+                var parameter = Database.CreateParameter("@ArticleId", query.ArticleID);
 
                 dbQuery.Parameters.Add(parameter);
