+++ Domain/Business/CommandServices/Comments/ICreateCommentCommandService.cs
@@ -1,0 +1,9 @@
+﻿using Domain.Data.Commands;
+using Domain.Data.Commands.Comments;
+
+namespace Domain.Business.CommandServices.Comments
+{
+    public interface ICreateCommentCommandService : ICommandService<CreateCommentCommand>
+    {
+    }
+}
